adc_phases: perform modulo with reference count, not demod count
The demodulation signal period count is divided (from the reference
count) by the scaling factor. Using this for the modulo can lose
precision. However, we only need the demodulation period count when
computing the real phase, so it's possible to get the best of both
worlds (integer modulo and no loss of precision).