replace build.rs with util
The sine and arctangent tables are not meant to change and therefore
don't need to be rebuilt by everyone compiling this code. Nonetheless,
the script is useful in case we want to change the size of the
array. Therefore, I've moved the table generation code to a utility
folder and am now tracking src/sin_table.txt and src/atan_table.txt.